requirements.txt
flask
flask-cors
numpy
tensorflow
pillow
opencv-python-headless
